refactor: 모집글, 찜 버튼 컴포넌트화
모집글, 찜 버튼 컴포넌트화
레이아웃 통일되게 수정
