+++ matrip/src/components/DibsBtn/index.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { styled } from 'styled-components';
+import { AiOutlineHeart } from 'react-icons/ai';
+
+function DibsBtn() {
+  return (
+    <DibsBtnBox>
+      <AiOutlineHeart size='24' />
+    </DibsBtnBox>
+  );
+}
+
+const DibsBtnBox = styled.button`
+  border: none;
+  background: none;
+  padding: 0 10px;
+`;
+
+export default DibsBtn;
