+++ matrip/src/pages/TripSchedule/TripScheduleStyle.tsx
@@ -107,6 +107,7 @@ export const userList = styled.div`
   display: flex;
   flex-wrap: wrap;
   justify-content: space-between;
+  gap: 15px;
 `;
 
 export const userListPostBox = styled.div`
