+++ matrip/src/pages/UserProfile/index.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import * as upc from './userProfileStyle';
+import * as gs from '../../styles/GlobalStyles';
 import { fdata } from '../../data/userdata';
 
 interface UserInfo {
@@ -15,26 +16,28 @@ const UserProfile: React.FC = () => {
   const prf: UserInfo = fdata[0];
 
   return (
-    <upc.MainContainer>
-      User Profile
-      <upc.ImgCtnr>
-        <upc.Img src={prf.imgurl} />
-      </upc.ImgCtnr>
-      <upc.IdCtnr>
-        <upc.IdText>{prf.id}</upc.IdText>
-      </upc.IdCtnr>
-      <upc.EmailCtnr>
-        <upc.EmailText>{prf.email}</upc.EmailText>
-      </upc.EmailCtnr>
-      <upc.BirthCtnr>
-        <upc.LabelName>생일</upc.LabelName>
-        <upc.BirthText>{prf.birth}</upc.BirthText>
-      </upc.BirthCtnr>
-      <upc.SexCtnr>
-        <upc.LabelName>성별</upc.LabelName>
-        <upc.SexText>{prf.sex}</upc.SexText>
-      </upc.SexCtnr>
-    </upc.MainContainer>
+    <gs.MainContainer>
+      <gs.MainBox>
+        User Profile
+        <upc.ImgCtnr>
+          <upc.Img src={prf.imgurl} />
+        </upc.ImgCtnr>
+        <upc.IdCtnr>
+          <upc.IdText>{prf.id}</upc.IdText>
+        </upc.IdCtnr>
+        <upc.EmailCtnr>
+          <upc.EmailText>{prf.email}</upc.EmailText>
+        </upc.EmailCtnr>
+        <upc.BirthCtnr>
+          <upc.LabelName>생일</upc.LabelName>
+          <upc.BirthText>{prf.birth}</upc.BirthText>
+        </upc.BirthCtnr>
+        <upc.SexCtnr>
+          <upc.LabelName>성별</upc.LabelName>
+          <upc.SexText>{prf.sex}</upc.SexText>
+        </upc.SexCtnr>
+      </gs.MainBox>
+    </gs.MainContainer>
   );
 };
